More disk robustness updates

+++ canine/localization/delocalization.py
@@ -123,24 +123,25 @@ def main(output_dir, jobId, patterns, copy, scratch):
             else:
                 print('INFO: matched output name "{0}" (pattern "{1}")'.format(name,  pattern), file = sys.stderr)
 
-    # compute checksums for all files
-    print('Computing CRC32C checksums ...', file = sys.stderr, flush = True, end = "")
-    pool = multiprocessing.Pool(8)
-    crc_results = []
-    for dest, f in matched_files:
-        crc_results.append((pool.apply_async(compute_crc32c, (f,)), dest, f))
+    # compute checksums for all files, if job exited successfully
+    if os.environ["CANINE_JOB_RC"] == "0":
+        print('Computing CRC32C checksums ...', file = sys.stderr, flush = True, end = "")
+        pool = multiprocessing.Pool(8)
+        crc_results = []
+        for dest, f in matched_files:
+            crc_results.append((pool.apply_async(compute_crc32c, (f,)), dest, f))
 
-    for res in crc_results:
-        crc = res[0].get()
-        dest = res[1]
-        f = res[2]
-        with open(os.path.join(
-            os.path.dirname(dest),
-            "." + os.path.basename(dest) + ".crc32c"
-          ), "w") as crc32c_file:
-            crc32c_file.write(crc + "\n")
-    pool.terminate()
-    print(' done', file = sys.stderr, flush = True)
+        for res in crc_results:
+            crc = res[0].get()
+            dest = res[1]
+            f = res[2]
+            with open(os.path.join(
+                os.path.dirname(dest),
+                "." + os.path.basename(dest) + ".crc32c"
+              ), "w") as crc32c_file:
+                crc32c_file.write(crc + "\n")
+        pool.terminate()
+        print(' done', file = sys.stderr, flush = True)
 
 if __name__ == '__main__':
     parser = argparse.ArgumentParser('canine-delocalizer')
